newt: add header file
