+++ pkg/urbit/include/vere/vere.h
@@ -3,6 +3,8 @@
 
 #include <uv.h>
 
+#include "vere/newt.h"
+
   /** Quasi-tunable parameters.
   **/
     /* First kernel this executable can boot.
@@ -37,79 +39,6 @@
         c3_w             pip_w;             //  target IPv4 address
         c3_s             por_s;             //  target port
       } u3_lane;
-
-    /* u3_moor_poke: poke callback function.
-    */
-      typedef c3_o (*u3_moor_poke)(void*, c3_d, c3_y*);
-
-    /* u3_moor_bail: bailout callback function.
-    */
-      typedef void (*u3_moor_bail)(void*, ssize_t err_i, const c3_c* err_c);
-
-    /* u3_meat: blob message block.
-    */
-      typedef struct _u3_meat {
-        struct _u3_meat* nex_u;
-        c3_d             len_d;
-        c3_y             hun_y[0];
-      } u3_meat;
-
-    /* u3_mess_type: in-process message type.
-    */
-      typedef enum {
-        u3_mess_head = 0,                   //  awaiting header
-        u3_mess_tail = 1                    //  awaiting body
-      } u3_mess_type;
-
-    /* u3_mess: blob message in process.
-    */
-      typedef struct _u3_mess {
-        u3_mess_type     sat_e;             //  msg type
-        union {                             //
-          struct {                          //  awaiting header:
-            c3_y         len_y[8];          //    header bytes
-            c3_y         has_y;             //    length
-          } hed_u;                          //
-          struct {                          //  awaiting body
-            u3_meat*     met_u;             //    partial message
-            c3_d         has_d;             //    length
-          } tal_u;                          //
-        };
-      } u3_mess;
-
-    /* u3_moat: inbound message stream.
-    */
-      typedef struct _u3_moat {
-        uv_pipe_t        pyp_u;             //  input stream
-        u3_moor_bail     bal_f;             //  error response function
-        void*            ptr_v;             //  callback pointer
-        u3_moor_poke     pok_f;             //  action function
-        u3_mess          mes_u;             //  message in progress
-        uv_timer_t       tim_u;             //  queue timer
-        u3_meat*         ent_u;             //  entry of message queue
-        u3_meat*         ext_u;             //  exit of message queue
-      } u3_moat;
-
-    /* u3_mojo: outbound message stream.
-    */
-      typedef struct _u3_mojo {
-        uv_pipe_t        pyp_u;             //  output stream
-        u3_moor_bail     bal_f;             //  error response function
-        void*            ptr_v;             //  callback pointer
-      } u3_mojo;
-
-    /* u3_moor: two-way message stream, linked list */
-      typedef struct _u3_moor {
-        uv_pipe_t        pyp_u;             //  duplex stream
-        u3_moor_bail     bal_f;             //  error response function
-        void*            ptr_v;             //  callback pointer
-        u3_moor_poke     pok_f;             //  action function
-        u3_mess          mes_u;             //  message in progress
-        uv_timer_t       tim_u;             //  queue timer
-        u3_meat*         ent_u;             //  entry of message queue
-        u3_meat*         ext_u;             //  exit of message queue
-        struct _u3_moor* nex_u;             //  next in list
-      } u3_moor;
 
     /* u3_dent: directory entry.
     */
@@ -1213,38 +1142,6 @@
       */
         u3_auto*
         u3_hind_io_init(u3_pier* pir_u);
-
-    /**  Stream messages.
-    **/
-      /* u3_newt_decode(): decode a (partial) length-prefixed byte buffer
-      */
-        void
-        u3_newt_decode(u3_moat* mot_u, c3_y* buf_y, c3_d len_d);
-
-      /* u3_newt_send(): write buffer to stream.
-      */
-        void
-        u3_newt_send(u3_mojo* moj_u, c3_d len_d, c3_y* byt_y);
-
-      /* u3_newt_read(): activate reading on input stream.
-      */
-        void
-        u3_newt_read(u3_moat* mot_u);
-
-      /* u3_newt_moat_info(); print status info.
-      */
-        void
-        u3_newt_moat_info(u3_moat* mot_u);
-
-      /* u3_newt_moat_stop(); newt stop/close input stream.
-      */
-        void
-        u3_newt_moat_stop(u3_moat* mot_u, u3_moor_bail bal_f);
-
-      /* u3_newt_mojo_stop(); newt stop/close output stream.
-      */
-        void
-        u3_newt_mojo_stop(u3_mojo* moj_u, u3_moor_bail bal_f);
 
     /** Mars interface.
     **/
